+++ src/main/java/com/rdsms/dao/IRdsmsDAO.java
@@ -13,16 +13,16 @@ public interface IRdsmsDAO {
 	
 	//Role
 	List<Role> getRoles();
-	Role getRoleByRoleId(int roleId);
+	Role getRoleByRoleId(Long roleId);
 	Role createRole(Role role);
-	Role updateRole(int roleId,Role role);
-	boolean deleteRole(int roleId);
+	Role updateRole(Long roleId,Role role);
+	boolean deleteRole(Long roleId);
 	
 	//User
 	List<User> getUsers();
 	User createUser(User user);
-	User updateUser(int userId, User user);
-	User getUserByUserId(int userId);
-	boolean deleteUser(int userId);
+	User updateUser(Long userId, User user);
+	User getUserByUserId(Long userId);
+	boolean deleteUser(Long userId);
 	
 }
